Revert "Merge branch 'master' of https://github.com/notrs/OMDataService into master"
This reverts commit fdc0b5829ede7ee3cdde689df7673874a5a024c9, reversing
changes made to 24bc0da782c0f4f89d3f7820a5e25608ddc54109.

+++ OMSDataService.DomainObjects/Models/ContractSearchResult.cs
@@ -88,7 +88,7 @@ namespace OMSDataService.DomainObjects.Models
         public decimal? ScheduledLessLoadoutQuantity { get; set; }
         public decimal? ScheduledLoads { get; set; }
         public decimal? WashedCancelledQuantity { get; set; }
-        public double? MzAdj { get; set; }
+        public decimal? MzAdj { get; set; }
 
         [NotMapped]
         public List<ContractGraphData> AppliedRemainingGraphData { get; set; }
